Organizing the project, adding navigation tools
