+++ App.js
@@ -1,43 +1,9 @@
 import React from 'react';
-import { StyleSheet,Text, View, Button } from 'react-native';
-import getQuote from './util/RandomQuoteApi';
-import Header from './components/header';
-import NewScreen from './components/NewScreen';
+import Quarantine_Time from './screens/Quarantine_Time';
+import Navigator from './routes/homeStack';
 
-export default class App extends React.Component {
-  
-  constructor() {
-    super();
-      this.state = {
-        loading: true,
-        quoteMeta: {},
-    }
-  }
-
-
-  async componentDidMount() {
-    this.setState({
-    quoteMeta: await getQuote(),
-    })
-  }
-
-  
-  render() {
+export default function App(){
   return (
-    <View style={styles.container}>
-       <Header />
-      <Button title="See Kanye's quote of the day." onPress={() => alert(this.state.quoteMeta.quote)}/>
-      <Button title = 'New Screen' onPress = {() => NewScreen}/>
-      <Button title = "About" onPress = {() => alert ("David, Matt, Nikolas, Justin ")}/>
-    </View>
+       <Navigator />
   );
-  }
-
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#FFEE58',
-  },
-});
