--- components/header 2.js
@@ -1,25 +1,0 @@
-import React from 'react';
-import { StyleSheet,Text, View } from 'react-native';
-
-export default function Header (){
-    return (
-        <View style = {styles.header}> 
-            <Text style = {styles.title}> Quarantine Time! </Text>
-        </View>
-    )
-}
-
-const styles = StyleSheet.create({
-    header: {
-        height: 80,
-        paddingTop: 38, 
-        backgroundColor: '#4E342E',
-        
-    },
-    title: {
-        textAlign: 'center',
-        color: '#FFF9C4',
-        fontSize: 20,
-        fontWeight: 'bold',
-    }
-});
